Generalize tf.test.compute_gradients to handle complex inputs and outputs
If inputs or outputs are complex, the Jacobian will still be real but the
corresponding Jacobian dimension(s) will be twice as large.  This is required
even if both input and output is complex since TensorFlow graphs are not
necessarily holomorphic, and may have gradients not expressible as complex
numbers.
Change: 115884848