Log an error in TensorBoard when we find out-of-order writes to event files.
This is unsupported and will break the event loader; the usual cause for this is
having multiple SummaryWriters, which is common enough that it's worth checking
for. This implements some crude checking that looks at the size of some of the
files in the directory whenever we switch from one file to the next.
Change: 120747280