Suppresses AssertionError when using InteractiveSession.
The following code would previously raise an AssertionError in the
garbage collector:

```python
sess = tf.InteractiveSession()
sess = tf.InteractiveSession()
```

This happens because the second session will be installed on the
default session stack before the first session is destroyed, which
leads to the assertion (of strict nesting) failing. Since this is
common behavior in IPython notebooks, we suppress this error for
InteractiveSession objects.

Fixes #2474.
Change: 123040755