TensorFlow: implement placement heuristic that takes generator nodes (nodes with one non-ref output and one consumer), and places it preferentially with its consumer.
For example:
    assign
   /      \
var       input

In the above graph, assign is bound to the device of 'var' due to the
reference edge.  This heuristic binds 'input' to the same device as
the assign, because it has only one consumer.

This addresses the general problem of colocating initializers with
their variables, and similar other cases.  There are very few reasons
to want to place the 'input' on a node other than its consumer (there
are some contrived cases, but that's why this is a heuristic).

This CL adds a test case for this small example above, illustrative
of the general problem.

An extension of this CL would be to do the same thing not just for
single output / single consumer nodes, but whenever all out edges of
a node connect to the same 'colcoation group'.
Change: 123896863