Actually fix the race condition for concurrent reading/writing of `tf.Graph`.
This change should fix issues where (e.g.) a node is added after queue
runners are started, and an error like `NotFoundError: FetchOutputs
node input_producer_Dequeue:0: not found` is reported.

The cause of the error was an incorrect multiple read of the
`Graph.version` property. A graph has a `version` property that is
(currently) defined to be the count of nodes in the graph. To keep the
front-end and session graphs in sync, the `tf.Session` maintains a
`_current_version` property, which is intended to be the last version
at which the graph was synced to the backend (via
TF_ExtendGraph). However, the logic was wrong in the face of
concurrent writes to the graph:

1. x = read `Session._current_version`.
2. y = read `Session._graph.version`.
3. If y > x:
   - Call `TF_ExtendGraph(Session._graph.as_graph_def(from_version=x))`
4. z = read `Session._graph.version`.
5. Set `Session._current_version` to z.

Written this way, the race condition is clear: a node added after step
3 and before step 4 would never be included in an TF_ExtendGraph call,
and so would never be added to the graph.

The fix involves returning the current graph version atomically with
the GraphDef, using a new internal method called
`Graph._as_graph_def()`. These are protected by the graph's mutex, to
ensure that no node can be added between creating the GraphDef and
reading the version.

Fixes #2957.
Change: 125356853