Bug fix: properly handle the case when the requested range cannot be satisfied.
When the range can't be satisfied, the GCS server returns the 416 HTTP code and
some additional information in the body. The curl library expects the exact amount of
bytes in the body as requested in the range, so curl_easy_peform fails with a
message like "Failed writing body (12 != 171)". With this fix, we don't check the
return value of curl_easy_peform if the HTTP code was 416.
Change: 125579779