Removing a confusing statement about training op (#3872)
* Correct Variable constructor docstring

It's a constructor, so it doesn't return anything. I removed the comment stating that it returns a Variable.

* Removing a confusing statement about training op

The statement seemed to suggest that optimizer.minimize() returns a Tensor, which it doesn't; I removed the statement.

To make the point even more explicit, I'm emphasizing that this is an op (maybe that's overkill though).