Implement ShapeManager, which will be used to perform graph-level shape inference in C++ using the C++ shape functions.
The interface for this class follows the way Graph building works: you
pass a Node at a time, and it creates an InferenceContext per Node.
On every call to AddNode, it requests the shapes from the
InferenceContexts of all the inputs, and it creates the
InferenceContext and runs the shape function in C++ to generate the
new node's InferenceContext data.  One possible use of this class
will be to perform a ReverseDFS (data-dep order) traversal of a Graph,
calling AddNode each time, while another use of this function will be
to add nodes as they are created in python, etc -- the only
requirement is that the nodes be added in the data dependency order,
so that a node's input shapes have been computed before the node.

As a start for doing as effectively as python does, we try to
materialize the input_tensor values by statically analyzing the graph
in the same way python's constant_value works, but I've only
implemented "Const" for now.  I'll do the rest in a follow up CL to
bring parity to the C++ implementation.

Since we do not know which inputs are needed before the shape function
runs, we run the inference function once to find out which tensors are
needed, and then we call it again once we materialize the tensors.

(I tried to get rid of the existing InferenceContext constructor,
 but it's difficult: moving the creation of shapes from strings
 relies on an InferenceContext in the first place! There's probably
 a way to do this, but this can be an internal refactor).

This is only a small step towards where we want to be with C++ Shape
inference.  This does not yet implement lazy-running of the graphs to
materialize as much of the graph as is computable, and does not yet
support iterative shape inference (forward and backward inference
until convergence).  These features will be built incrementally as we
discover any subtle nuances that make these features challenging.
Change: 130981592