Add to the C++ Node class the ability to fetch input nodes and edges by index.  There are various locations in code where users currently use iteration to find the edge by its already known index, and these functions would be useful to accomplish. In addition, this implements the equivalent functionality of 'op.inputs[i]' in our python Operation class.
Given the new functionality, it exposed a weird use of NoOp for nodes
that actually had multiple inputs.  Modified the test to use custom op
definitions to be more correct.

Currently this iterates over the edge list, which in the common case
will be fast and introduces no additional state to Node.  In the future
we may want to revisit this.
Change: 131299794