Load the `tf.contrib` module on demand only.
This change makes it so that `tf.contrib` and the libraries it
contains are not loaded until their first use. The rationale for this
change is that `tf.contrib` has a lower code quality than the other
`tf.*` libraries, and it frequently breaks users due to undeclared
dependencies or other churn.

See e.g. https://github.com/tensorflow/tensorflow/issues/2388 for an
example of where `import tensorflow as tf` fails due to a breakage in
`tf.contrib`. This issue
(https://github.com/tensorflow/tensorflow/issues/2154) shows a
dependency on a particular version of pandas.
Change: 131332657