Add ProjectorScatterPlotAdapter, which turns 'hover, selected, neighbors' into color / scale / opacity / label arrays to power the scatter plot.
I started moving the trace code to be array-driven like the rest of the render attributes, but it turns out that the easiest way to do that is to rewrite the trace visualizer as a generic "LineSegmentVisualizer" that accumulates all line segments into a single render call. Reverted but coming soon.
Change: 136351181