Add a new op split_v that can handle variable size splits.
Aside from being useful on its own, this op also makes the implementation
of the gradient of concat much more efficient.  Previously a new slice op was
created in the graph for every input tensor to concat.  This op moves that
logic inside of one op.  The overhead could be quite significant in cases
with 100+ input Tensors to concat.
Change: 138822942