pip.sh: change the semantics of environment-variable switch NO_TEST_ON_INSTALL to make it more useful
Previously, if NO_TEST_ON_INSTALL is set to 1, it'll skip all the following: python unit tests on install, user ops test, tutorial tests and ffmpeg integration test. But most frequently, we want to be able to skip the python unit tests on install, but still run user ops tests, tutorial tests and/or ffmpeg integration test. This change enables that.
Change: 142820160