Improve the performance of resize_bilinear.
The previous implementation of resize_bilinear operation is a straight
forward, naive implementation. Unfortunately, it leaves a significant amount
of room for improvement.

This change improves the performance of the operation by avoiding unnecessary
recomputation (for where the operation was CPU bound), and avoids redundant
memory references (for where the operation was memory bound).

In my benchmark (loosely based on inception image pipeline), this implementation
change improves speed between 1.5-2x.

Additionally, to ensure correctness, I've preserved the old implementation of
resize_bilinear, and added a number of tests that initalize a random image,
and ensure that the outputs are identical.
Change: 144726607