Improve error message when converting unconvertable lists of objects to Tensor.
At the moment, we don't know how to convert a list of objects such as
TensorShape or Dimension to a Tensor, or a list of objects where
type promotion may be slightly ambiguous (e.g., [1.0, Dimension(1)]).
Before this change, the code would dump an unhelp error message when
trying to interpret these objects using compat.as_bytes(), so instead
we catch the error and raise a slightly more informative error that suggests
where the original list comes from to the user.
Change: 151653212