Clarify and correct Android Sections in WORKSPACE (#8689)
* Clarify and correct Android Sections

Build Tools version updated to latest version shipping with Android Studio.
Added some comments to clarify both this and NDK Level details that have been causing confusion.

* Update Build Tools Version

Android studio uses 25.0.2 now (I think) and complains about incorrect bulid tools version in gradle. This fixes it. Not sure if this is unique to my android studio though because I have latest version?

* Requested changes to WORKSPACE for Android

Clarification of comments to the android section at the top. Comments added to avoid common confusion.

* Update Dockerfile.android to clarify build tools.

I don't understand this dockerfile too much but I gather that this will download and use the correct build_tools_version regardless? Nonetheless, updated to 25.0.2 as a minimum and added a comment to assist troubleshooters in the future.

* Update build_tools_version and add comment

Added comment to clarify build_tools_version updates and changed it to 25.0.2

* Update WORKSPACE

* Update WORKSPACE

* Update Dockerfile.android

* Update WORKSPACE

* Updated for "the", lines and line wrapping