Make TensorShape and PartialTensorShape share memory representation
Once we use a custom NodeInfo class to represent NodeDef in memory, shape attrs can be represented using a native C++ class.  Unfortunately, NodeInfo won't know whether a shape attr is supposed to be partial or not.  After this change, both classes have the same memory representation as their common base class TensorShapeRep.  TensorShapeRep is always safely static_castable to PartialTensorShape, and can be safely cast to TensorShape as long as num_elements() != -1.

As a benefit, users of PartialTensorShape will be faster.

RELNOTES: Unify memory representations of TensorShape and PartialTensorShape.  As a consequence, tensors now have a maximum of 254 dimensions, not 255.
PiperOrigin-RevId: 156628598