Optionally use new session API in Python API
The new session API is enabled if ops._USE_C_API is True, otherwise the original behavior is maintained (the default). Most of the tests in session_test.py are now run with the new API. The new session API runs the TF_Graph that is built alongside the Python graph when ops._USE_C_API is True.

PiperOrigin-RevId: 157272139