Introduce tensorboard_zip_file build rule
This rule can depend on web_library or tensorboard_html_binary. In
both cases it will create a .zip file containing all the transitive
web server paths. This can be used to deploy static assets to web
servers.

A small change was also made to Vulcanize to support path overriding.

PiperOrigin-RevId: 157655047