Disable writing of compressed checkpoints.
Snappy compression (and decompression) was enabled after the
1.1 release (in commit 31b3f00b3cc4cf4e940ca8b8da2896fcf5deccf8).
This means that checkpoints produced by the 1.2.0 release candidates
will cause TensorFlow 1.1 (and prior) binaries to crash as they
CHECK fail when trying to load snappy-compressed tables.

To ease transition, disable writing of compressed checkpoints in
1.2.0 for now.

Reconsider this in the next release.

PiperOrigin-RevId: 157675189