Fix a bug in importing for colocation's interaction with udf device fns.
When a user imports a graph_def within a device function, the current
code properly propagates the colocation attribute to a node that
is colocated with another.  However, a user's device function still
independently applies, causing the system to fail at runtime
because the colocation attribute doesn't match the device function
it is also given.

To address this, we add another phase to importing after we know all
user device functions have applied.  In this phase, in which we only
process ops that are colocated with others, we fetch the colocation
root's device, and apply that to the local node's device.  If it
is possible that the two devices merge cleanly, we attempt to
merge the devices so the user's specification still applies, as
long as it maintains the colocation property.

PiperOrigin-RevId: 158551722