[Tensorboard] Add a dashboard for profile tooling plugin.
A new tab in the index page will be added when both frontend and backend (plugin) are
ready.

o The dashboard only supports one tool at this point i.e. trace viewer via iframe.
o In the profile dashboard, a "run" represents a profile run and the "tag" indicates the tool to use for visualization.

PiperOrigin-RevId: 158823605