Refactor pip.sh into independent jobs
pip.sh consists of multiple tasks, such as
* testing pip package on a clean virtualenv install
* running all bazel_pip tests in a "non-clean" virtualenv
* testing user ops
* testing tfdbg binaries
etc.

Previously they run in a way in which if an earlier task fails, none of the following tasks will run. This CL corrects that by separating the code into multiple tasks and ensure that they always run regardless of the status of preceding tasks.

Also adds color to the FAIL / PASS messages.

PiperOrigin-RevId: 162807396