Creates a ProcessFunctionLibraryRuntime object that manages / owns all the FunctionLibraryRuntime (FLR) objects, one per device.
This change makes sure that we create FLR objects for each device, regardless of whether we have that device in the pruned graph. This is necessary for cross device (intra-process) function execution in which its possible to have a graph only present on one device (post-pruning) that wants to execute a function on another device.

PiperOrigin-RevId: 164180464