Allow partial shape inference for `tf.nn.conv2d_transpose`
This fix tries to address the issue raised in 8972 where
it was not possible to infer a partial shape for
`tf.nn.conv2d_transpose`. In case any dimensions are not
defined, the shape will be totally undefined.

This fix utilizes `tensor_util.constant_value_as_shape`
so that partial shape is possible.

This fix fixes 8972.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>