Extend the summing methods used in tf.metrics.auc (currently a trapezoidal Riemman sum) by a minoring Rieman sum defined locally as "the left Riemann sum if the curve is locally decreasing and the right Riemann sum if the curve is locally increasing" and a majoring Rieman sum (the opposite).
For monotone intervals, the minoring summation method results to a lower bound of the real AUC while the majoring summation method leads to an upper bound of the real AUC.

The AUC-PR of a model always predicting 0.0 would be 0.5 with 'trapezoidal' sum, 0.0 with 'minoring' sum and 1.0 with 'majoring' sum.

Computing the delta between 'minoring' and 'majoring' AUC provides a confidence metric on the empirical estimation.

PiperOrigin-RevId: 170326074