Copy public tf.contrib.graph_editor.reroute_{inputs,outputs} docs.
There are multiple references "see reroute_inputs" which are unhelpful because the full docstring now only exists on _reroute_sgv_inputs (likewise for reroute_outputs). Copy most of the docstring to reroute_{inputs,outputs} so that it is outputted in the docs.

Update some other dangling doc references from _reroute to _reroute_sgv, but that docstring will not be included the docs.

PiperOrigin-RevId: 171821659