[XLA] Merge large parameter-shaped tuples into their users in DOT graphs.
It's common to have a while loop whose body computation has one
parameter, a giant tuple.  Then we have to draw edges from that tuple to
a bunch of get-tuple-element nodes, which are used throughout the while
loop's body.  This results in many long, difficult-to-follow edges.

In practice, the big tuple really functions as N separate parameters.
This patch represents it this way visually, erasing the big tuple and
replacing it with the get-tuple-element users, which we style like
parameters.

Future work is figuring out how to do something similar for the tuple op
at the bottom of while loop bodies.  This will be harder, because it
will require breaking the invariant that every HLO corresponds to zero
or one nodes in the dot graph.

PiperOrigin-RevId: 173584100