Implements specifying default approximations for layer_collection.
Currently, the default approximation to use for each layer is hard-coded as
a default argument to each registration function. This CL instead specifies
these default values as properties which the user can modify. Additionally,
the user can identify groups of linked parameters that should always use a
specified approximation when registered. This should make it easier for
users to experiment with different approximations.

PiperOrigin-RevId: 175955141