Call Graph._add_op in Operation.__init__ (and remove existing calls).
Without this change, ops manually constructed via Operation.__init__
must be passed to Graph._add_op to keep the graph in a consistent
state. Failure to do so is particularly disasterous with the C API
enabled, as more Operation methods rely on Graph._nodes_by_name, which
is updated in Graph._add_op (e.g. Operation.inputs will fail if the
inputs have not been added to the graph).

An alternative to this change is to require that all
Operation.__init__ callers also call Graph._add_op (we don't currently
do this in ops_test.py, although I imagine all non-test callers
do). While this is effectively the current contract, it forces callers
of Operation.__init__, which is a public API, to use _add_op, which is
private.

One downside of this change is that it will break existing
Graph._add_op calls, since the op will already have been
added. However, _add_op is a private API.

PiperOrigin-RevId: 176180386