Fix edge case with ImportGraphDefOption.uniquify_names = true.
This change fixes the case where a newly-generated uniquified name
conflicts with another NodeDef being imported (the original NodeDef
names are required to be unique among each other, so this is only an
issue when we create new names).

Note that this behavior is not well defined in the Python
import_graph_def method. It will always generate unique names, but the
exact naming scheme may depend on the order the NodeDefs are
imported. I didn't write a corresponding Python unit test or try to
make this change produce the same names for this reason.

PiperOrigin-RevId: 177872720