WhileContext.AddBackpropIndexedSlicesAccumulator shouldn't create invalid shapes
AddBackpropIndexedSlicesAccumulator effectively creates a while loop
where one of the loop variables changes shape on every iteration (it
concats together the accumulated values). This means the while loop
body should not propagate the input tensor shape, which is what this
change acheives. The current code works because
Operation._update_input doesn't check the new shape, but it will check
with the C API enabled.

PiperOrigin-RevId: 178278910