Prototype for serializing the Checkpointable graph.
Includes protocol buffers which will likely be added to BundleHeaderProto. We'll pass this protocol buffer plus the standard map of variable names to variable objects to the Saver, both of which will then be stored along with the checkpoint. This extra information means that the path we stored to a variable in its checkpoint name can be removed as long as there is another in the checkpointed graph.

PiperOrigin-RevId: 178946669