[TF Eager] Non-public API for registering TF functions with the graph.
We have tfe.defun, which registers and calls the function (possibly
with caching).  But for ops that accept a function as an attribute,
we need to register/attach the function based on an input signature,
and query the output signature, but not actually create an op containing
the function call.

PiperOrigin-RevId: 179230747