Checkpointable: Move the checkpoint-grouping utility out of the unit test file
Renames Saver -> CheckpointableSaver in preparation for exposing the necessary
symbols in tf.contrib.eager.

There's a pending change for Optimizers, and Asim is handling Layers/Model. Once
those are checked in, we should be able to save/restore everything in the eager
examples (or at least the mnist one...). Still plenty more to make
Checkpointable, but it should be usable at that point.

PiperOrigin-RevId: 187221803